# backend/studio-manager-service/src/main/resources/openapi/AgentStudio-Config-API.yaml
swagger: '2.0'
info:
  version: v1
  title: config manager API
  description: config manager API
schemes:
  - https
host: localhost.com
basePath: /
...
paths:
...
  /v1/{project_id}/agent-manager/structured-messages:
    post:
      tags:
        - MessageManagement
      summary: 添加结构化信息
      description: 添加结构化信息
      operationId: CreateStructuredMessages
      produces:
        - application/json
      parameters:
        - name: project_id
          in: path
          required: true
          type: string
          maxLength: 64
          description: 租户项目id
          pattern: ^[a-zA-Z0-9_-]+$
        - name: workspace_id
          in: query
          required: true
          type: string
          description: 项目空间id
          pattern: ^[a-zA-Z0-9_()\-]+$
          maxLength: 64
          minLength: 1
        - name: structured_messages_params
          in: body
          required: true
          description: 结构化信息参数设置
          schema:
            $ref: '#/definitions/StructuredInfoRequest'
      responses:
        '200':
          description: 返回的结构化信息
          schema:
            $ref: '#/definitions/StructMessage'
        '400':
          description: Bad Request 请求错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '401':
          description: Unauthorized 鉴权失败
          schema:
            type: string
        '403':
          description: Forbidden 没有操作权限
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '404':
          description: Not Found 找不到资源
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '500':
          description: Internal Server Error 服务内部错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
      x-order: 1
    put:
      tags:
        - MessageManagement
      summary: 修改结构化信息
      description: 修改指定ID的结构化信息
      operationId: UpdateStructuredMessages
      produces:
        - application/json
      parameters:
        - name: project_id
          in: path
          required: true
          type: string
          maxLength: 64
          description: 租户项目id
          pattern: ^[a-zA-Z0-9_-]+$
        - name: workspace_id
          in: query
          required: true
          type: string
          description: 项目空间id
          pattern: ^[a-zA-Z0-9_()\-]+$
          maxLength: 64
          minLength: 1
        - name: structured_messages_params
          in: body
          required: true
          description: 结构化信息更新参数
          schema:
            $ref: '#/definitions/StructuredInfoRequest'
      responses:
        '200':
          description: 更新成功的结构化信息
          schema:
            $ref: '#/definitions/StructMessage'
        '400':
          description: Bad Request 请求错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '401':
          description: Unauthorized 鉴权失败
          schema:
            type: string
        '403':
          description: Forbidden 没有操作权限
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '404':
          description: Not Found 找不到指定资源
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '500':
          description: Internal Server Error 服务内部错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
      x-order: 0
    get:
      tags:
        - MessageManagement
      summary: 查询结构化信息列表
      description: 查询当前项目空间下的所有结构化信息
      operationId: ListStructuredMessages
      produces:
        - application/json
      parameters:
        - name: project_id
          in: path
          required: true
          type: string
          maxLength: 64
          description: 租户项目id
          pattern: ^[a-zA-Z0-9_-]+$
        - name: workspace_id
          in: query
          required: true
          type: string
          description: 项目空间id
          pattern: ^[a-zA-Z0-9_()\-]+$
          maxLength: 64
          minLength: 1
        - name: offset
          in: query
          required: false
          description: 分页记录的起始位置偏移量,默认值0
          type: integer
          format: int32
          default: 0
          minimum: 0
          maximum: 100000
        - name: limit
          in: query
          required: false
          description: 每一页的数量,默认10
          type: integer
          format: int32
          default: 1000
          minimum: 1
          maximum: 1000
        - name: sort
          in: query
          required: false
          description: 排序顺序，默认降序
          type: string
          maxLength: 5
          default: desc
        - name: name
          in: query
          required: false
          description: 结构化消息名称
          type: string
          maxLength: 192
          pattern: '^.{0,64}$'
        - name: category
          in: query
          required: false
          description: 结构化消息类型
          type: string
          maxLength: 16
        - name: import_method
          in: query
          required: false
          description: 结构化信息导入方式
          type: string
          maxLength: 16
        - name: status
          in: query
          required: false
          description: 状态信息
          type: string
          maxLength: 32
        - name: creator
          in: query
          required: false
          description: 创建人
          type: string
          maxLength: 32
        - name: updater
          in: query
          required: false
          description: 修改人
          type: string
          maxLength: 32
      responses:
        '200':
          description: 结构化信息列表
          schema:
            $ref: '#/definitions/StructMessageListRsp'
        '400':
          description: Bad Request 请求错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '401':
          description: Unauthorized 鉴权失败
          schema:
            type: string
        '403':
          description: Forbidden 没有操作权限
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '404':
          description: Not Found 找不到资源
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '500':
          description: Internal Server Error 服务内部错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
      x-order: 3
    delete:
      tags:
        - MessageManagement
      summary: 批量删除结构化信息
      description: 根据提供的信息ID列表批量删除结构化信息
      operationId: DeleteStructuredMessages
      produces:
        - application/json
      parameters:
        - name: project_id
          in: path
          required: true
          type: string
          maxLength: 64
          description: 租户项目ID
          pattern: ^[a-zA-Z0-9_-]+$
        - name: workspace_id
          in: query
          required: true
          type: string
          description: 项目空间id
          pattern: ^[a-zA-Z0-9_()\-]+$
          maxLength: 64
          minLength: 1
        - name: delete_request
          in: body
          required: true
          description: 结构化信息删除请求参数
          schema:
            type: array
            minItems: 1
            maxItems: 100
            items:
              $ref: '#/definitions/StructuredInfoRequestDelete'
      responses:
        '200':
          description: 批量删除操作结果
          schema:
            $ref: '#/definitions/BatchDeleteRsp'
        '400':
          description: Bad Request 请求参数错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '401':
          description: Unauthorized 鉴权失败
          schema:
            type: string
        '403':
          description: Forbidden 没有操作权限
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '404':
          description: Not Found 资源不存在
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '500':
          description: Internal Server Error 服务内部错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
      x-order: 2
  /v1/{project_id}/agent-manager/structured-messages/upload-file:
    post:
      tags:
        - MessageManagement
      summary: 上传消息模板
      description: 上传消息模板
      operationId: uploadStructuredMessages
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: project_id
          in: path
          required: true
          type: string
          maxLength: 64
          description: 租户项目id
          pattern: ^[a-zA-Z0-9_-]+$
        - name: workspace_id
          in: query
          required: true
          type: string
          description: 项目空间id
          pattern: ^[a-zA-Z0-9_()\-]+$
          maxLength: 64
          minLength: 1
        - name: file
          in: formData
          required: true
          description: 导入结构化信息的文件
          type: file
      responses:
        '200':
          description: 结构化信息列表
          schema:
            $ref: '#/definitions/StructuredMessagesUploadResult'
        '400':
          description: Bad Request 请求错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '401':
          description: Unauthorized 鉴权失败
          schema:
            type: string
        '403':
          description: Forbidden 没有操作权限
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '404':
          description: Not Found 找不到资源
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '500':
          description: Internal Server Error 服务内部错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
      x-order: 3
  /v1/{project_id}/agent-manager/structured-messages/export-file:
    post:
      tags:
        - MessageManagement
      summary: 导出消息模板
      description: 上传消息模板
      operationId: exportStructuredMessages
      produces:
        - application/json
      parameters:
        - name: project_id
          in: path
          required: true
          type: string
          maxLength: 64
          description: 租户项目id
          pattern: ^[a-zA-Z0-9_-]+$
        - name: workspace_id
          in: query
          required: true
          type: string
          description: 项目空间id
          pattern: ^[a-zA-Z0-9_()\-]+$
          maxLength: 64
          minLength: 1
        - name: export_params
          in: body
          required: true
          description: 导出参数设置
          schema:
            $ref: '#/definitions/ExportMessagesParams'
      responses:
        '200':
          description: 导出文档
          schema:
            type: file
        '400':
          description: Bad Request 请求错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '401':
          description: Unauthorized 鉴权失败
          schema:
            type: string
        '403':
          description: Forbidden 没有操作权限
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '404':
          description: Not Found 找不到资源
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '500':
          description: Internal Server Error 服务内部错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
      x-order: 4
  /v1/{project_id}/agent-manager/structured-messages/export-template:
    post:
      tags:
        - MessageManagement
      summary: 结构化信息模板
      description: 结构化信息模板获取
      operationId: ExportStructuredTemplate
      produces:
        - application/json
      parameters:
        - name: project_id
          in: path
          required: true
          type: string
          maxLength: 64
          description: 租户项目id
          pattern: ^[a-zA-Z0-9_-]+$
        - name: workspace_id
          in: query
          required: true
          type: string
          description: 项目空间id
          pattern: ^[a-zA-Z0-9_()\-]+$
          maxLength: 64
          minLength: 1
      responses:
        '200':
          description: 导出Excel
          schema:
            type: file
        '400':
          description: Bad Request 请求错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '401':
          description: Unauthorized 鉴权失败
          schema:
            type: string
        '403':
          description: Forbidden 没有操作权限
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '404':
          description: Not Found 找不到资源
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
        '500':
          description: Internal Server Error 服务内部错误
          schema:
            $ref: 'AgentStudio-Manager-API.yaml#/definitions/ErrorRsp'
      x-order: 5
definitions:
  ErrorRsp:
    type: object
    description: 完整错误信息
    properties:
      error_code:
        type: string
        description: 错误码
      error_msg:
        type: string
        description: 错误信息
      details:
        type: array
        items:
          $ref: '#/definitions/ErrorDetail'
        description: 调用接口返回的错误信息
      error_reason:
        type: string
        description: 错误原因
      error_suggestion:
        type: string
        description: 错误建议
  ErrorDetail:
    type: object
    description: 错误信息
    properties:
      error_code:
        type: string
        description: 错误码
      error_msg:
        type: string
        description: 错误信息
...
  BatchDeleteRsp:
    type: object
    description: 删除数据源响应体
    required:
      - count
    properties:
      count:
        type: integer
        format: int64
        description: 总数
      ids:
        type: array
        description: 删除的数据源id列表
        minItems: 1
        maxItems: 100
        items:
          pattern: ^[a-zA-Z0-9_-]+$
          minLength: 1
          maxLength: 64
          type: string
          description: 数据源列表
  StructMessageListRsp:
    type: object
    description: 查询结构化信息列表返回体
    required:
      - count
    properties:
      count:
        type: integer
        format: int64
        description: 总数
      message_list:
        type: array
        items:
          $ref: '#/definitions/StructMessage'
        description: 结构化信息列表
  StructuredInfoRequest:
    type: object
    description: 结构化信息请求
    properties:
      id:
        type: string
        description: 结构体消息ID（创建时可选，编辑时必填）
        maxLength: 64
        pattern: ^[a-zA-Z0-9_-]+$
      name:
        type: string
        description: 结构体消息名称（必填）
        example: 无权限
        minLength: 2
        maxLength: 64
      category:
        type: string
        description: 结构体消息分类（可选）
        example: 异常
        maxLength: 64
      content:
        type: object
        description: 结构体消息体（可选）
        additionalProperties: true
        maxProperties: 10
      visibility:
        type: string
        description: 结构体可见范围（可选）
        enum:
          - personal
          - tenant
          - workspace
    required:
      - name
      - visibility
  StructuredInfoRequestDelete:
    type: object
    description: 删除结构化信息请求
    properties:
      id:
        type: string
        description: 结构体消息ID（创建时可选，编辑时必填）
        maxLength: 64
        pattern: ^[a-zA-Z0-9_-]+$
      name:
        type: string
        description: 结构体消息名称（必填）
        example: 无权限
        minLength: 2
        maxLength: 64
      category:
        type: string
        description: 结构体消息分类（可选）
        example: 异常
        maxLength: 64
      content:
        type: object
        description: 结构体消息体（可选）
        additionalProperties: true
        maxProperties: 10
      visibility:
        type: string
        description: 结构体可见范围（可选）
        enum:
          - personal
          - tenant
          - workspace
    required:
      - id
  StructMessage:
    type: object
    description: 结构化信息返回实体
    properties:
      id:
        type: string
        description: 结构体消息唯一标识符
        pattern: ^[a-zA-Z0-9_-]+$
        maxLength: 64
      name:
        type: string
        description: 结构体消息名称
        maxLength: 500
      category:
        type: string
        description: 结构体消息分类
        maxLength: 500
      content:
        type: object
        description: 结构体消息内容主体
        additionalProperties: true
      visibility:
        type: string
        description: 结构体可见范围
        maxLength: 500